refactor: (#793) roles 변수명 수정 및 @JsonFormat 추가
